+++ openhands/sdk/conversation/base.py
@@ -1,9 +1,9 @@
-from abc import ABC, abstractmethod
-from pathlib import Path
-from typing import TYPE_CHECKING, Any, Protocol
+from abc import abstractmethod
+from typing import TYPE_CHECKING, Protocol
 
 from openhands.sdk.conversation.conversation_stats import ConversationStats
 from openhands.sdk.conversation.secrets_manager import SecretValue
+from openhands.sdk.conversation.system_mixins.base import SystemMixin
 from openhands.sdk.conversation.types import ConversationID
 from openhands.sdk.llm.message import Message
 from openhands.sdk.security.confirmation_policy import (
@@ -47,7 +47,7 @@ class ConversationStateProtocol(Protocol):
         ...
 
 
-class BaseConversation(ABC):
+class BaseConversation(SystemMixin):
     @property
     @abstractmethod
     def id(self) -> ConversationID: ...
@@ -86,84 +86,3 @@ class BaseConversation(ABC):
 
     @abstractmethod
     def close(self) -> None: ...
-
-    # System-level operations
-    @abstractmethod
-    def execute_bash(
-        self,
-        command: str,
-        cwd: str | Path | None = None,
-        timeout: float = 30.0,
-    ) -> dict[str, Any]:
-        """Execute a bash command on the system.
-
-        These system operations are independent of the conversation itself.
-        A system can spawn multiple conversations, and these functionalities
-        can be accessed without connecting to a system. The reason these
-        methods are tied to conversation is mostly for convenience, and maybe
-        later they will be scoped based on the workspace of the conversation.
-
-        Args:
-            command: The bash command to execute
-            cwd: Working directory for the command (optional)
-            timeout: Timeout in seconds (defaults to 30.0)
-
-        Returns:
-            dict: Result containing stdout, stderr, exit_code, and other metadata
-
-        Raises:
-            Exception: If command execution fails
-        """
-        ...
-
-    @abstractmethod
-    def file_upload(
-        self,
-        source_path: str | Path,
-        destination_path: str | Path,
-    ) -> dict[str, Any]:
-        """Upload a file to the system.
-
-        These system operations are independent of the conversation itself.
-        A system can spawn multiple conversations, and these functionalities
-        can be accessed without connecting to a system. The reason these
-        methods are tied to conversation is mostly for convenience, and maybe
-        later they will be scoped based on the workspace of the conversation.
-
-        Args:
-            source_path: Path to the source file
-            destination_path: Path where the file should be uploaded
-
-        Returns:
-            dict: Result containing success status and metadata
-
-        Raises:
-            Exception: If file upload fails
-        """
-        ...
-
-    @abstractmethod
-    def file_download(
-        self,
-        source_path: str | Path,
-        destination_path: str | Path,
-    ) -> dict[str, Any]:
-        """Download a file from the system.
-
-        These system operations are independent of the conversation itself.
-        A system can spawn multiple conversations, and these functionalities
-        can be accessed without connecting to a system. The reason these
-        methods are tied to conversation is mostly for convenience, and maybe
-        later they will be scoped based on the workspace of the conversation.
-
-        Args:
-            source_path: Path to the source file on the system
-            destination_path: Path where the file should be downloaded
-
-        Returns:
-            dict: Result containing success status and metadata
-
-        Raises:
-            Exception: If file download fails
-        """
-        ...
